Modify _is_package_available to skip version check unless version is requested
This is to work-around a problem with discovering the availability of CuPy - which fails if the user has installed CuPy pre-built wheels (e.g. `pip install cupy-cuda12x`) rather than from source (`pip install cupy`). The reason this function fails in this case is because it assumes that the name of importable package matches the name of the library package. This is not always true in the case of CuPy (as already described) but also breaks in the case of several other libraries (e.g. `faiss` and `onnxruntime`) to name a couple.

The solution implemented here is to just skip the version check unless the version information has been explicitly requested by the caller. This should mean that any callers that ask for the version (of which there actually currently aren't any) continue to work as before, and any callers that don't ask for the version will be more forgiving.